backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
transformers
torch
google-generativeai
pydantic
httpx
python-multipart
alembic
aiosqlite
psycopg2-binary
